GPS Sensor
GPS Sensor is currently fully operational. The code is set up to have the Longitude and Latitude values sent over UART from GPS sensor to the STM32L496 microcontroller. The STMs receives the data in a buffer over UART protocol and I have mapped char points to the position in the buffer and also including N,S,W,E tags for polarity.